Refactor resampling logic into its own class
